requirements.txt
numpy==2.2.0
pandas==2.2.3
python-dotenv==1.0.1
pytz==2024.2
pyupbit==0.2.34
pyzmq==26.2.0
slack_sdk==3.33.5